requirements.txt
boto3>=1.26.0
huggingface-hub>=0.16.4
tqdm>=4.65.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.3.1
python-dotenv>=0.21.0
torchsummary>=1.5.1
numpy>=1.21.0
requests>=2.26.0
pyarrow>=14.0.1